fix: createdAt null 수정

+++ BackEnd/Review/src/main/java/com/sshmarket/review/application/dto/ReviewRatingAndNum.java
@@ -13,6 +13,6 @@ public class ReviewRatingAndNum {
 
     private Long articleId;
     private Double starRating;
-    private Integer reviewCnt;
+    private Long reviewCnt;
 
 }
